Fixed bugs pertaining to byte buffer
more needed to work

+++ SequencerLib/ISequencerConfiguration.h
@@ -7,7 +7,7 @@ namespace SequencerLib
 	{
 		virtual int GetRecurrenceIntervalPerSecond(void) = 0;
 		virtual int GetErrorCorrectiveCoeffitient(void) = 0;
-		virtual int GetResolutionPerSecond(void) = 0;
+		virtual float GetResolutionPerSecond(void) = 0;
 		virtual UINT64 GetBufferSize(void) = 0;
 		virtual UINT64 GetSmallestPointBetweenBeats(void) = 0;
 		virtual int GetTotalBeats(void) = 0;
